班次人员管理完善

+++ Application/巡检信息/frmDefectType.cs
@@ -70,14 +70,14 @@ namespace WorkStation
             {
                 foreach (TreeListNode node in treeList1.Nodes)
                 {
-                    ids += treeVisitor(node);
+                   ids += treeVisitor(node);
                 }
             }
             else
             {
                 if (treeList1.FocusedNode != null)
                 {
-                    ids += treeVisitor(treeList1.FocusedNode);
+                   ids += treeVisitor(treeList1.FocusedNode);
                 }
             }
             if (ids.Trim() != "")
@@ -102,11 +102,10 @@ namespace WorkStation
         private string treeVisitor(TreeListNode areaNode)
         {
             string id = "";
-            id += areaNode.GetDisplayText("ID") + ",";
-            
+            id += areaNode.GetDisplayText("ID") + ",";            
             foreach (TreeListNode n in areaNode.Nodes)
             {
-                 id += treeVisitor(n);
+               id += treeVisitor(n);
             }
             return id;
         }
@@ -177,7 +176,7 @@ namespace WorkStation
         //启/停用编辑缺陷类型
         private void barCheckItem2_CheckedChanged(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-           // treeList1.OptionsBehavior.Editable = barCheckItem2.Checked;
+           //treeList1.OptionsBehavior.Editable = barCheckItem2.Checked;
             frmDefectTypeNew typenew = new frmDefectTypeNew();
             typenew.IsEdit = true;
             typenew.DefectTypeID = treeList1.FocusedNode.GetDisplayText("DefectType_ID");
@@ -189,7 +188,7 @@ namespace WorkStation
         //启/停用编辑缺陷
         private void barCheckItem3_CheckedChanged(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-           // gridView1.OptionsBehavior.Editable = barCheckItem3.Checked;
+           //gridView1.OptionsBehavior.Editable = barCheckItem3.Checked;
             frmDefectNew fectnew = new frmDefectNew();
             fectnew.IsEdit = true;
             fectnew.DefectTypeID = treeList1.FocusedNode.GetDisplayText("DefectType_ID");
@@ -201,25 +200,32 @@ namespace WorkStation
         //删除缺陷类型
         private void barButtonItem3_ItemClick(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
-            if (treeList1.FocusedNode == null) return;
-            string del = "delete from defecttype where id="+treeList1.FocusedNode.GetDisplayText("ID");
-            if (SqlHelper.ExecuteNonQuery(del) != 1)
+            if(treeList1.FocusedNode == null) return;
+            if (MessageBox.Show("您确定删除该信息吗？", "提示信息", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)
             {
-                MessageBox.Show("删除失败，稍后再试！");
+                string del = "delete from defecttype where id=" + treeList1.FocusedNode.GetDisplayText("ID");
+                if (SqlHelper.ExecuteNonQuery(del) != 1)
+                {
+                    MessageBox.Show("删除失败，稍后再试！");
+                }
+                BindTreeList();
             }
-            BindTreeList();
         }
     
         //删除缺陷
         private void barButtonItem4_ItemClick(object sender, DevExpress.XtraBars.ItemClickEventArgs e)
         {
             if (gridView1.FocusedRowHandle < 0) return;
-            string del = "delete from defect where id=" + gridView1.GetRowCellValue(gridView1.FocusedRowHandle, "ID");
-            if (SqlHelper.ExecuteNonQuery(del) != 1)
+            if (MessageBox.Show("您确定删除该信息吗？", "提示信息", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == DialogResult.OK)
             {
-                MessageBox.Show("删除失败，稍后再试！");
+                string del = "delete from defect where id=" + gridView1.GetRowCellValue(gridView1.FocusedRowHandle, "ID");
+
+                if (SqlHelper.ExecuteNonQuery(del) != 1)
+                {
+                    MessageBox.Show("删除失败，稍后再试！");
+                }
+                BindDgv();
             }
-            BindDgv();
         }
 
         //编辑缺陷类型
